+++ test/basic_freexgraph_test.py
@@ -85,7 +85,7 @@ def test_get_not_existing(valid_basic_execution_graph):
     assert node is None
 
 
-def test_visitation(valid_basic_execution_graph, visitor_test):
+def test_visitation(valid_basic_execution_graph, visitor_test, node_test_class):
     #
     #
     #     ,_____, id2
@@ -104,6 +104,32 @@ def test_visitation(valid_basic_execution_graph, visitor_test):
     assert visitor_test.visited[2].startswith("id4_")
     assert visitor_test.visited[3].startswith("id3_")
     assert visitor_test.visited[4].startswith("id5_")
+
+    # check that modification of the node
+    id5 = visitor_test.visited[4]
+    node_5 = valid_basic_execution_graph.get_node(id5)
+    assert node_5.metadata is None
+    assert node_5.fork_id is None
+    assert sorted(node_5.parents) == sorted(
+        {visitor_test.visited[3], visitor_test.visited[2]}
+    )
+
+    valid_basic_execution_graph.replace_node(
+        node_test_class(
+            id5,
+            metadata="ChocoboForTheWin",
+            parents={"ThisIdWouldMessUpEverything"},
+            graph_ref=None,
+            fork_id="Chocobo",
+        )
+    )
+
+    node_5 = valid_basic_execution_graph.get_node(id5)
+    assert node_5.metadata is "ChocoboForTheWin"
+    assert node_5.fork_id is None
+    assert sorted(node_5.parents) == sorted(
+        {visitor_test.visited[3], visitor_test.visited[2]}
+    )
 
 
 # used only in test_visitation_custom_hook and never somewhere else